removed if-else statement in step function; split sensors into abstract class and obstacle and goal sensor; bug fix of resetfunction for resetLimits; reset example script to old version; sensor flag defines if sensors are added
