+++ planarenvs/pointRobot/envs/pointRobotEnv.py
@@ -42,8 +42,8 @@ class PointRobotEnv(PlanarEnv):
         self._limUpVel = self._limits['vel']['high']
         self._limUpAcc = self._limits['acc']['high']
         self._limUpFor = self._limits['for']['high']
-        self.observation_space['x'] = spaces.Box(low=-self._limUpPos, high=self._limUpPos, dtype=np.float64)
-        self.observation_space['xdot'] = spaces.Box(low=-self._limUpVel, high=self._limUpVel, dtype=np.float64)
+        self.observation_space.spaces['x'] = spaces.Box(low=-self._limUpPos, high=self._limUpPos, dtype=np.float64)
+        self.observation_space.spaces['xdot'] = spaces.Box(low=-self._limUpVel, high=self._limUpVel, dtype=np.float64)
 
 
     @abstractmethod
